fixing a hole where the rain gets in
